Restructure of files
